requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.8.3
python-multipart==0.0.9
